requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Computer Vision and OCR
opencv-python-headless==4.8.1.78
numpy==1.24.4
pillow==10.1.0

# PaddleOCR (optional)
paddleocr==2.7.3
paddlepaddle==2.5.2

# Document generation
reportlab==4.0.7
fpdf2==2.7.6
openpyxl==3.1.2
pandas==2.1.4

# Storage (optional S3 support)
boto3==1.34.0
botocore==1.34.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0

# Security
cryptography==41.0.8

# Optional: Additional QR/Barcode support
pyzbar==0.1.9
pylibdmtx==0.1.10